Fixed an issue with showing images for recognize_duration
Added images to the library, but those are there just for the sake of
tracking changes in future. The code has been changed to use the base64
encoded versions of those that are part of the JS file. This way the
location of the images is of no importance. If something is ever changed
or added, then these changes should be consistently instorduced ithin
the JS file.